requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
langgraph==0.2.34
streamlit==1.39.0
httpx==0.27.2
faiss-cpu==1.8.0
sentence-transformers==3.0.1
numpy==1.26.4
